Refactor to TypeScript
