+++ .gitmodules
@@ -1,0 +1,3 @@
+[submodule "vendor/alfa"]
+	path = vendor/alfa
+	url = https://github.com/Siteimprove/alfa.git
