+++ bin/tsconfig.json
@@ -1,0 +1,7 @@
+{
+  "extends": "../tsconfig.json",
+  "files": [
+    "a11y-check.ts"
+  ],
+  "references": [{ "path": "../lib" }]
+}
