+++ lib/constants/index.ts
@@ -1,7 +1,6 @@
-const OUTCOME = {
+export const OUTCOME = {
   CANNOT_TELL: 'cantTell',
   FAILED: 'failed',
   INAPPLICABLE: 'inapplicable',
   PASSED: 'passed',
 };
-exports.OUTCOME = OUTCOME;
