+++ lib/tsconfig.json
@@ -1,0 +1,15 @@
+{
+  "extends": "../tsconfig.json",
+  "files": [
+    "constants/index.ts",
+    "index.ts",
+    "tasks/audit.ts",
+    "tasks/qna.ts",
+    "tasks/report.ts",
+    "tasks/save.ts",
+    "tasks/scrape.ts",
+    "util/log.ts",
+    "util/package.ts"
+  ],
+  "references": [{ "path": "../vendor/alfa/packages" }]
+}
