+++ lib/util/log.ts
@@ -1,8 +1,6 @@
-function log(message) {
+export default function log(message) {
   const prefix = `[a11y-check]:`;
   const messageStr = message.toString();
 
   console.log(`${prefix} ${messageStr}`);
 }
-
-module.exports = log;
