+++ package.json
@@ -2,37 +2,25 @@
   "name": "@accessibility/accessibility-checker",
   "version": "0.0.0",
   "description": "Check Accessibility by the help of act-rules and @siteimprove/alfa",
-  "main": "./bin/check-accessibility.js",
+  "main": "lib/index.js",
   "repository": "git@github.com:Accessibility-Foundation/accessibility-checker.git",
   "author": "Roel Antonisse <roel.antonisse@gmail.com>",
   "license": "MIT",
   "private": true,
   "bin": {
-    "a11y-check": "./bin/a11y-check.js"
-  },
-  "scripts": {
-    "preinstall": "npm install git-dependency --no-shrinkwrap && npm run git-modules:install && npm run alfa:build",
-    "install": "npm run alfa:link",
-    "alfa:update": "npm run git-modules:install && npm run alfa:build && npm run alfa:link",
-    "alfa:build": "yarn install --cwd ./git_modules/siteimprove-alfa",
-    "alfa:link": "yarn install --no-lockfile --ignore-scripts",
-    "git-modules:install": "npx git-dependency install",
-    "a11y-check": "node ./bin/a11y-check.js"
+    "a11y-check": "bin/a11y-check.js"
   },
   "workspaces": [
-    "git_modules/siteimprove-alfa/packages/*"
+    "vendor/alfa/packages/*"
   ],
-  "directories": {
-    "gitDependencies": "./git_modules"
-  },
-  "gitDependencies": {
-    "siteimprove-alfa": "https://github.com/Bartimeus-Accessibility/alfa.git"
-  },
-  "devDependencies": {
-    "git-dependency": "0.0.3"
+  "scripts": {
+    "build": "tsc --build"
   },
   "dependencies": {
     "commander": "^2.19.0",
     "prompt-async": "^0.9.9"
+  },
+  "devDependencies": {
+    "typescript": "^3.4.5"
   }
 }
